% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{any_of}
\alias{any_of}
\title{anyof combinator, takes a vector of characters and returns a parser that
mactches any of them}
\usage{
any_of(chars)
}
\arguments{
\item{chars}{}
}
\description{
anyof combinator, takes a vector of characters and returns a parser that
mactches any of them
}
