% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{then}
\alias{then}
\title{then
the 'and then' combinator
'}
\usage{
then(parserL, parserR)
}
\arguments{
\item{parserR}{}
}
\description{
then
the 'and then' combinator
'
}
\examples{

pchar('p') \%then\% pchar('a')('parse')
}
